[simplify] Implement besselsimp.
In reference to issue 2917.

This commit adds (the start of) a routine for simplifying bessel-type functions.
It currently only works on besseli and besselj and only implements a very
simplistic algorithm. It however works nicely in many cases occurring in practice.

Add tests for this.

Also adapt some existing tests (mostly transforms) to the new, better simplification.